Remove superfluous (duplicated) Message types.
